refactor: task numeration
