+++ src/main/kotlin/homeworks/homework1/task2/eratosthenesSieve/EratosthenesSieve.kt
@@ -1,4 +1,4 @@
-package homeworks.homework1.task1.eratosthenesSieve
+package homeworks.homework1.task2.eratosthenesSieve
 
 interface EratosthenesSieve {
     fun getPrimesUpToTheBoundary(bound: Int): List<Int>
