+++ src/main/kotlin/homeworks/homework1/task2/eratosthenesSieve/EratosthenesSieveImperative.kt
@@ -1,4 +1,4 @@
-package homeworks.homework1.task1.eratosthenesSieve
+package homeworks.homework1.task2.eratosthenesSieve
 
 import kotlin.math.sqrt
