+++ src/main/kotlin/homeworks/homework1/task2/eratosthenesSieve/InvalidBoundException.kt
@@ -1,4 +1,4 @@
-package homeworks.homework1.task1.eratosthenesSieve
+package homeworks.homework1.task2.eratosthenesSieve
 
 class InvalidBoundException(bound: Int) :
     IllegalArgumentException("passed bound=$bound must be greater then 0")
